[Perf] Change pdbClose to test if files are locked before inducing GCs
The original code triggers 6 full GCs, the time is a multiple of the
size of the binary (since all of the IL/metadata structures are still on
the heap at this point). This takes ~600ms on my test project.

Changed to perform a rudimentary check if the file is locked before
inducing GC. This is very hard to test - will run with it and observe.